update uRWELL semaring

+++ src/programs/Simulation/mcsmear/CGEMSmearer.h
@@ -13,7 +13,7 @@ class cgem_config_t
 		
 
 	double m_CGEM_WORK_FUNCTION;
-	int m_CGEM_FANO_FACTOR; 
+	double m_CGEM_FANO_FACTOR; 
 	double m_CGEM_ENERGY_THRES;
 	double m_CGEM_SPATIAL_RESOLUTION;
 	double m_CGEM_TIMING_RESOLUTION;
